+++ HISTORY.md
@@ -1,2 +1,7 @@
+Version ```0.0.2``` - ```2021.02.08```
+    - added program to generate file with channels for openHab3 MQTT generic thing
+    - added timestamp as a topic
+    - code cleaning
+
 Version ```0.0.1``` - ```2021.02.07```
     - initial version